瓶の底判定の修正
