+++ 外部変形/binaryEnv/binary_calc.py
@@ -43,3 +43,23 @@ def calc_area(frame):
   for i in range(len(frame)):
       area += (frame[i].x * frame[0].y - frame[i].y * frame[0].x) / 2 if i == len(frame) - 1 else (frame[i].x * frame[i + 1].y - frame[i].y * frame[i + 1].x) / 2
   return abs(area)
+
+# 直線AB上の点Pから垂直に落とした点Hを求める
+def Get_vertical_intersection(A, B, P):
+    """直線AB上の点Pから垂直に落とした点を求める
+
+    Args:
+    A (Point): _探索軸の始点
+    B (Point): _探索軸の終点
+    P (Point): _AとBの二点のリスト
+
+    Returns:
+    point_on_judge_line (Point): _判定軸上の点
+    """
+    AB = Point(B.x - A.x, B.y - A.y)
+    AP = Point(P.x - A.x, P.y - A.y)
+
+    k = Point.dot(AB, AP) / (AB.magnitude() ** 2)
+    OH = Point(k * AB.x, k * AB.y)
+    point_on_judge_line =  Point(OH.x + A.x, OH.y + A.y)
+    return point_on_judge_line
